Move tidy and Sphinx functionality to wbc module
